feat: add `Database` class
The Database class abstracts the database access operations related to
fetching data with queries.

Per this implementation, a database client type was introduced in
`database.ts`. In the future, if there ever needs to be a new database
to access data, the client can be added to this type. The fetch function
should operate just the same.

This database class makes the assumption that our database schema will
not change, or at the very least, stay consistent for each database
operation.

Abstracting away this functionality separates the view layer ever more
from the data, which is great, since we are already abstracting away the
data fetching through an api facility (using SvelteKit's `/api` route).